Rename and rescope things
