+++ apps/prairielearn/src/components/AssessmentBadge.tsx
@@ -1,4 +1,4 @@
-import { renderHtml } from '@prairielearn/preact/client';
+import { renderHtml } from '@prairielearn/preact';
 
 import { type AssessmentInstanceUrlParts, getAssessmentInstanceUrl } from '../lib/client/url.js';
