+++ apps/prairielearn/src/components/GitHubButton.tsx
@@ -1,4 +1,4 @@
-import { renderHtml } from '@prairielearn/preact/client';
+import { renderHtml } from '@prairielearn/preact';
 
 export function GitHubButton({ gitHubLink }: { gitHubLink: string | null }) {
   if (!gitHubLink) return null;
