+++ apps/prairielearn/src/components/IssueBadge.tsx
@@ -1,6 +1,6 @@
 import clsx from 'clsx';
 
-import { renderHtml } from '@prairielearn/preact/client';
+import { renderHtml } from '@prairielearn/preact';
 
 export function IssueBadge({
   count,
