+++ apps/prairielearn/src/ee/pages/instructorAiGenerateDraftEditor/instructorAiGenerateDraftEditor.html.tsx
@@ -1,6 +1,6 @@
 import { html, unsafeHtml } from '@prairielearn/html';
-import { Hydrate } from '@prairielearn/preact';
-import { renderHtml } from '@prairielearn/preact/client';
+import { Hydrate } from '@prairielearn/preact/server';
+import { renderHtml } from '@prairielearn/preact';
 import { run } from '@prairielearn/run';
 
 import { HeadContents } from '../../../components/HeadContents.js';
