+++ apps/prairielearn/src/middlewares/authzHasCoursePreviewOrInstanceView.tsx
@@ -2,7 +2,7 @@ import { type Request, type Response } from 'express';
 import asyncHandler from 'express-async-handler';
 
 import * as error from '@prairielearn/error';
-import { Hydrate } from '@prairielearn/preact';
+import { Hydrate } from '@prairielearn/preact/server';
 
 import { PageLayout } from '../components/PageLayout.js';
 import { getPageContext } from '../lib/client/page-context.js';
