+++ apps/prairielearn/src/pages/instructorAssessmentQuestions/instructorAssessmentQuestions.html.tsx
@@ -1,4 +1,4 @@
-import { Hydrate } from '@prairielearn/preact';
+import { Hydrate } from '@prairielearn/preact/server';
 
 import { PageLayout } from '../../components/PageLayout.js';
 import { AssessmentSyncErrorsAndWarnings } from '../../components/SyncErrorsAndWarnings.js';
