+++ apps/prairielearn/src/pages/instructorAssessmentStatistics/instructorAssessmentStatistics.html.tsx
@@ -2,7 +2,7 @@ import _ from 'lodash';
 import { z } from 'zod';
 
 import { html } from '@prairielearn/html';
-import { renderHtml } from '@prairielearn/preact/client';
+import { renderHtml } from '@prairielearn/preact';
 
 import { PageLayout } from '../../components/PageLayout.js';
 import { AssessmentSyncErrorsAndWarnings } from '../../components/SyncErrorsAndWarnings.js';
