+++ apps/prairielearn/src/pages/instructorAssessments/instructorAssessments.html.tsx
@@ -1,7 +1,7 @@
 import { EncodedData } from '@prairielearn/browser-utils';
 import { formatInterval } from '@prairielearn/formatter';
 import { html } from '@prairielearn/html';
-import { renderHtml } from '@prairielearn/preact/client';
+import { renderHtml } from '@prairielearn/preact';
 import { run } from '@prairielearn/run';
 
 import { AssessmentModuleHeading } from '../../components/AssessmentModuleHeading.js';
