+++ apps/prairielearn/src/pages/instructorCourseAdminModules/instructorCourseAdminModules.html.tsx
@@ -1,5 +1,5 @@
 import { html } from '@prairielearn/html';
-import { renderHtml } from '@prairielearn/preact/client';
+import { renderHtml } from '@prairielearn/preact';
 
 import { AssessmentModuleHeading } from '../../components/AssessmentModuleHeading.js';
 import { PageLayout } from '../../components/PageLayout.js';
