+++ apps/prairielearn/src/pages/instructorCourseAdminSharing/instructorCourseAdminSharing.html.tsx
@@ -1,7 +1,7 @@
 import { z } from 'zod';
 
 import { type HtmlSafeString, escapeHtml, html } from '@prairielearn/html';
-import { renderHtml } from '@prairielearn/preact/client';
+import { renderHtml } from '@prairielearn/preact';
 
 import { Modal } from '../../components/Modal.js';
 import { PageLayout } from '../../components/PageLayout.js';
