+++ apps/prairielearn/src/pages/instructorInstanceAdminAccess/instructorInstanceAdminAccess.html.tsx
@@ -1,6 +1,6 @@
 import { formatDate } from '@prairielearn/formatter';
 import { html } from '@prairielearn/html';
-import { renderHtml } from '@prairielearn/preact/client';
+import { renderHtml } from '@prairielearn/preact';
 
 import { CommentPopover } from '../../components/CommentPopover.js';
 import { PageLayout } from '../../components/PageLayout.js';
